Implemented the confirm/cancel feature for reservation statuses in dashboard
