+++ app/controllers/vaccines_controller.rb
@@ -43,22 +43,22 @@ class VaccinesController < ApplicationController
     redirect_to vaccines_path, notice: "Vaccine was successfully deleted"
   end
 
-  def dashboard
-    # Vaccines you have listed
-    @vaccines = policy_scope(Vaccine).where(user: current_user)
+  # def dashboard
+  #   # Vaccines you have listed
+  #   @vaccines = policy_scope(Vaccine).where(user: current_user)
 
-    # Vaccines you have reserved(booked)
-    @reservations = policy_scope(Reservation).where(user: current_user)
+  #   # Vaccines you have reserved(booked)
+  #   @reservations = policy_scope(Reservation).where(user: current_user)
 
-    # Reservations for your vaccines
-    # Find all reservations where the owner of the vaccine is the current_user
-    @reservations_for_seller = []
-    @vaccines.each do |vaccine|
-      vaccine.reservations.each do |reservation|
-        @reservations_for_seller << reservation
-      end
-    end
-  end
+  #   # Reservations for your vaccines
+  #   # Find all reservations where the owner of the vaccine is the current_user
+  #   @reservations_for_seller = []
+  #   @vaccines.each do |vaccine|
+  #     vaccine.reservations.each do |reservation|
+  #       @reservations_for_seller << reservation
+  #     end
+  #   end
+  # end
 
   private
