+++ app/views/dashboard/dashboard.html.erb
@@ -1,0 +1,15 @@
+<div class="container">
+
+  <div class="container" id='your-vaccines'>
+    <%= render 'shared/vaccines-table' %>
+  </div>
+
+  <div class="container" id='your-vaccines'>
+    <%= render 'shared/reservations-for-seller-table' %>
+  </div>
+
+  <div class="container" id="your-reservations">
+    <%= render 'shared/reservations-table' %>
+  </div>
+
+</div>
