+++ app/views/vaccines/show.html.erb
@@ -25,9 +25,16 @@
     <!-- TO ADD => Owner of this vaccine listing can see all reservations -->
 
     <!-- Currently anyone can see the reservations -->
-    <% @vaccine.reservations.each do |reservation| %>
+<!--     <% @vaccine.reservations.each do |reservation| %>
       <% if reservation.status == 'Pending' %>
         <p> <%= reservation.user.email %> has a reservation for this vaccine at this date: <%= reservation.date %>.</p>
+      <% end %>
+    <% end %> -->
+    <% if @vaccine.user == current_user %>
+      <% current_user.reservations.each do |reservation| %>
+        <% if reservation.status == 'Pending' %>
+          <p> <%= reservation.user.email %> has a reservation for this vaccine at this date: <%= reservation.date %>.</p>
+        <% end %>
       <% end %>
     <% end %>
   </div>
