+++ config/routes.rb
@@ -4,8 +4,8 @@ Rails.application.routes.draw do
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
   resources :vaccines do
     resources :reservations
-    patch "dashboard", to: 'reservations#confirm', as: 'reservation_confirm'
-    patch "dashboard", to: 'reservations#cancel', as: 'reservation_cancel'
   end
-  get 'dashboard', to: 'vaccines#dashboard', as: 'dashboard'
+  get 'dashboard', to: 'dashboard#dashboard', as: 'dashboard'
+  patch "reservations/:id/confirm", to: 'dashboard#confirm', as: 'reservation_confirm'
+  patch "reservations/:id/cancel", to: 'dashboard#cancel', as: 'reservation_cancel'
 end
